requirements.txt
pandas>=1.5
numpy>=1.24
scikit-learn>=1.2
matplotlib>=3.6
pytest>=7.2
requests>=2.28
python-dotenv>=1.0
flask>=2.2
flask-socketio>=5.3
scipy>=1.10
pyotp>=2.8
websocket-client>=1.5
pytest-cov
black
flake8
mypy
pyyaml>=6.0
toml>=0.10.2
